Merge scopes
Summary:
Block scoping mandate several new scopes to be created while compiling JS code.
These extra scopes would become ```Environments``` (created with
```CreateInnerEnvironment```) at runtime, and would lead to performance regressions
in block scope-enabled code.

This diff introduces the ```ScopeMerger``` optimization pass. The idea is that all the
scopes created during IRGen could be merged, if doing so is non-observable by a running
JS program.

Thus, this optimization will always merge scopes (within a function) unless the scope is
```dynamic``` and any of its variables is captured. A ```dynamic``` scope is one created by
a loop body. This is needed to allow closures nested inside of the loop body to capture
variables without "sharing" the captures with other closures. For example, in the following
excerpt

```
  closures = []
  for (const [init, step] of foo()) {
    let i = init;
    closures.push(() => { i+= step; print(i) });
  }
```

it is imperative that each closure keep its own copy of ```i```.

This optimization allows block scoping-enabled code to perform identically to
es5 code in most situations, and to be more efficient than babel lowered
code when variables in ```dynamic``` scopes are captured (ES5 lowering in this
case would introduce a new closure, and thus incur the cost of a function call. See
https://babeljs.io/docs/babel-plugin-transform-block-scoping#throwifclosurerequired).

Reviewed By: mhorowitz

Differential Revision: D39738560

fbshipit-source-id: 356d549824d238c52fbbbe930e2384fc773ea199