Make default behavior of OwnKeyFlags explicit
Summary:
Add an option for IncludeEnumerable in OwnKeyFlags. Remove the implicit
default behavior of always including enumerable properties. This way,
we have the capability of requesting only the nonenumerable keys,
something which was impossible before this diff. The current use-case
for this is handling nonenumerable property shadowing.

For now, proxy does not support excluding enumerable keys. If support
is needed for this it can be added later.

Reviewed By: jpporto

Differential Revision: D43645421

fbshipit-source-id: 17700f2c90ffe56ee33b842adcb05cb9bce029ed